+++ supvisors/tests/test_viewsupvisors.py
@@ -66,27 +66,48 @@ def test_write_navigation(mocker, view):
     assert mocked_nav.call_args_list == [call(mocked_root)]
 
 
-def test_write_header(mocker, view):
+def test_write_header(mocker, supvisors, view):
     """ Test the write_header method. """
     # patch context
     mocker.patch.object(view.sup_ctx, 'get_state_modes',
                         return_value={'fsm_statename': SupvisorsStates.DEPLOYMENT.name,
                                       'starting_jobs': True, 'stopping_jobs': False})
+    mocked_super = mocker.patch('supvisors.web.viewhandler.ViewHandler.write_header')
     mocked_actions = mocker.patch.object(view, 'write_supvisors_actions')
     # build root structure
     state_mid = create_element()
     starting_mid = create_element()
     stopping_mid = create_element()
-    mocked_root = create_element({'state_mid': state_mid, 'starting_mid': starting_mid, 'stopping_mid': stopping_mid})
-    # test call with no auto-refresh
-    view.write_header(mocked_root)
-    assert mocked_root.findmeld.call_args_list == [call('state_mid'), call('starting_mid'), call('stopping_mid')]
+    master_mid = create_element()
+    mocked_header = create_element({'state_mid': state_mid, 'master_name_mid': master_mid,
+                                    'starting_mid': starting_mid, 'stopping_mid': stopping_mid})
+    # test call with no master
+    view.write_header(mocked_header)
+    assert mocked_super.call_args_list == [call(mocked_header)]
+    assert mocked_header.findmeld.call_args_list == [call('state_mid'), call('starting_mid'), call('stopping_mid'),
+                                                     call('master_name_mid')]
     assert state_mid.content.call_args_list == [call('DEPLOYMENT')]
     assert starting_mid.attrib['class'] == 'blink'
     assert not starting_mid.replace.called
     assert stopping_mid.attrib['class'] == ''
     assert stopping_mid.replace.call_args_list == [call('')]
-    assert mocked_actions.call_args_list == [call(mocked_root)]
+    assert master_mid.content.call_args_list == [call('none')]
+    assert mocked_actions.call_args_list == [call(mocked_header)]
+    mocked_header.reset_all()
+    mocker.resetall()
+    # test call with master
+    supvisors.context.local_status.state_modes.master_identifier = '10.0.0.1:65000'
+    view.write_header(mocked_header)
+    assert mocked_super.call_args_list == [call(mocked_header)]
+    assert mocked_header.findmeld.call_args_list == [call('state_mid'), call('starting_mid'), call('stopping_mid'),
+                                                     call('master_name_mid')]
+    assert state_mid.content.call_args_list == [call('DEPLOYMENT')]
+    assert starting_mid.attrib['class'] == 'blink'
+    assert not starting_mid.replace.called
+    assert stopping_mid.attrib['class'] == ''
+    assert stopping_mid.replace.call_args_list == [call('')]
+    assert master_mid.content.call_args_list == [call('10.0.0.1')]
+    assert mocked_actions.call_args_list == [call(mocked_header)]
 
 
 def test_write_supvisors_actions(view):